+++ src/types/code-editor.ts
@@ -2,8 +2,20 @@
  * Supported Languages
  */
 export const enum CodeLanguage {
+  C = 'c',
+  CPP = 'cpp',
+  CSHARP = 'csharp',
+  GO = 'go',
+  JAVA = 'java',
   JAVASCRIPT = 'javascript',
   JSON = 'json',
+  MYSQL = 'mysql',
+  PHP = 'php',
+  PGSQL = 'pgsql',
+  PYTHON = 'python',
+  RUBY = 'ruby',
+  SQL = 'sql',
+  TYPESCRIPT = 'typescript',
 }
 
 /**